+++ src/report/cpu.rs
@@ -4,13 +4,13 @@ use serde::Serialize;
 #[derive(Debug, Serialize)]
 pub struct CpuForm {
     pub job_id: Option<String>,
-    pub n: usize,
+    pub n: u64,
     pub times: Vec<f64>,
 }
 
 impl CpuForm {
     #[inline]
-    pub fn new(job_id: Option<String>, n: usize, times: Vec<f64>) -> Self {
+    pub fn new(job_id: Option<String>, n: u64, times: Vec<f64>) -> Self {
         Self { job_id, n, times }
     }
 }
